Map_location
We can find that most of the heritages are located in the south of Scotland, as well as in the north and East. Also, there is a small distribution in southern and central England. It is worth noting that from the red dot, many relics are concentrated at the border between southern Scotland and England and the coastal areas, which is likely due to the demand for brick Trade, so the factories are built in these areas. We can also find that Glasgow has a good location and can be a transportation hub. Perhaps those brick factories has contributed to its becoming the largest city in Scotland.